+++ src/main/java/dynamo/RidesDao.java
@@ -1,17 +1,31 @@
 package dynamo;
 
 import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
 
 import com.amazonaws.regions.Regions;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
 import com.amazonaws.services.dynamodbv2.document.DynamoDB;
+import com.amazonaws.services.dynamodbv2.document.Item;
+import com.amazonaws.services.dynamodbv2.document.PrimaryKey;
+import com.amazonaws.services.dynamodbv2.document.PutItemOutcome;
 import com.amazonaws.services.dynamodbv2.document.Table;
+import com.amazonaws.services.dynamodbv2.document.UpdateItemOutcome;
+import com.amazonaws.services.dynamodbv2.document.spec.DeleteItemSpec;
+import com.amazonaws.services.dynamodbv2.document.spec.GetItemSpec;
+import com.amazonaws.services.dynamodbv2.document.spec.UpdateItemSpec;
+import com.amazonaws.services.dynamodbv2.document.utils.ValueMap;
 import com.amazonaws.services.dynamodbv2.model.AttributeDefinition;
+import com.amazonaws.services.dynamodbv2.model.AttributeValue;
 import com.amazonaws.services.dynamodbv2.model.KeySchemaElement;
 import com.amazonaws.services.dynamodbv2.model.KeyType;
 import com.amazonaws.services.dynamodbv2.model.ProvisionedThroughput;
+import com.amazonaws.services.dynamodbv2.model.ReturnValue;
 import com.amazonaws.services.dynamodbv2.model.ScalarAttributeType;
+import com.amazonaws.services.dynamodbv2.model.ScanRequest;
+import com.amazonaws.services.dynamodbv2.model.ScanResult;
 
 import models.ClientRequest;
 
@@ -20,83 +34,18 @@ import models.ClientRequest;
  */
 public class RidesDao {
 
-    private AmazonDynamoDB dbClient;
-    private DynamoDB dynamoDB;
+	private AmazonDynamoDB dbClient;
+	private DynamoDB dynamoDB;
 
-//    Table table = dynamoDB.getTable("Movies");   // Name of the database table
-    
-    
-    // Needs to deal with the credentials...
-    public RidesDao() {
-    	
-        dbClient = AmazonDynamoDBClientBuilder.standard()
-                        .withRegion(Regions.US_WEST_1)
-                        .build();
-        dynamoDB = new DynamoDB(dbClient);     
-        /*Ryan's Region
-        dbClient = AmazonDynamoDBClientBuilder.standard()
-                .withRegion(Regions.US_EAST_2)
-                .build();
-        dynamoDB = new DynamoDB(dbClient);   
-        */
-    }
 
-    
-    
-    
-    
-    // Create Table 
-    public void createTable() {
-        dbClient = AmazonDynamoDBClientBuilder.standard()
-                .withRegion(Regions.US_WEST_1)
-                .build();
-dynamoDB = new DynamoDB(dbClient);   
-    	/*
-        dbClient = AmazonDynamoDBClientBuilder.standard()   // Client
-                .withRegion(Regions.US_EAST_2)
-                .build();
-        dynamoDB = new DynamoDB(dbClient);   
-        */
-        String tableName = "RideTable2"; //Name for Table
+	// Needs to deal with the credentials...
+	public RidesDao() {
 
-        try {
-            System.out.println("Attempting to create table; please wait...");
-            Table table = dynamoDB.createTable(tableName,
-                Arrays.asList(new KeySchemaElement("Email", KeyType.HASH), // Partition
-                                                                          // key
-                    new KeySchemaElement("Name", KeyType.RANGE)), // Sort key
-                Arrays.asList(new AttributeDefinition("Email", ScalarAttributeType.S),
-                    new AttributeDefinition("Name", ScalarAttributeType.S)),
-                new ProvisionedThroughput(10L, 10L));
-            table.waitForActive();
-            System.out.println("Success.  Table status: " + table.getDescription().getTableStatus());
+		dbClient = AmazonDynamoDBClientBuilder.standard().withRegion(Regions.US_WEST_1).build();
+		dynamoDB = new DynamoDB(dbClient);
 
-        }
-        catch (Exception e) {
-            System.err.println("Unable to create table: ");
-            System.err.println(e.getMessage());
-        }
-    }
-    
-    
-    //Delete Table
-    public void deleteTable(String tableName) {
-        Table table = dynamoDB.getTable(tableName);
-        try {
-            System.out.println("Issuing DeleteTable request for " + tableName);
-            table.delete();
+	}
 
-            System.out.println("Waiting for " + tableName + " to be deleted...this may take a while...");
-
-            table.waitForDelete();
-        }
-        catch (Exception e) {
-            System.err.println("DeleteTable request failed for " + tableName);
-            System.err.println(e.getMessage());
-        }
-    }
-    
-    
     // Todo: Implement insert, update, read, readAll, delete
     public void insert(String email, String name, int year, String phoneNumber, String church, boolean attendance)
     {
