+++ src/test/java/Playground.java
@@ -7,6 +7,8 @@ public class Playground {
 
     public static void main(String[] args) {
         RidesDao dao = new RidesDao();
+
+
         //dao.insert();
         System.out.println("hello");
     }
